Tile, Entity, rudimenti di controlli e character
Implementazione di tile ed entity, prima idea di un sistema di controlli e personaggio

+++ SmashSDL2/Vector2f.h
@@ -3,15 +3,15 @@
 struct Vector2f {
 
 	Vector2f()
-		:x(0.0), y(0.0)
+		:x(0), y(0)
 	{}
 	
 	Vector2f(int X, int Y)
 		:x(X), y(Y)
 	{}
 
-	double x;
+	int x;
 
-	double y;
+	int y;
 
 };
